refactor: Make the server a library for easier testing
